requirements.txt
# Core dependencies
typing>=3.7.4
stockfish>=3.28.0
python-chess>=1.10.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
